+++ src/main/java/duke/task/Task.java
@@ -1,7 +1,6 @@
 package duke.task;
 
 import java.time.DateTimeException;
-import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
 import duke.Duke;
@@ -51,7 +50,7 @@ abstract class Task {
         }
     }
 
-    private static TaskType getTaskType(String input) {
+    protected static TaskType getTaskType(String input) {
         if (input.startsWith("todo")) {
             return TaskType.TODO;
         }
@@ -67,7 +66,7 @@ abstract class Task {
         return null;
     }
 
-    private static String squareBracketWrapper(String input) {
+    protected static String squareBracketWrapper(String input) {
         return "[" + input + "]";
     }
 
@@ -89,160 +88,5 @@ abstract class Task {
         TODO,
         DEADLINE,
         EVENT
-    }
-
-    private static final class TodoTask extends Task {
-        public TodoTask(String task) throws Duke.WrongFormatException {
-            String description = getDescription(task);
-            if (description == null) {
-                throw new Duke.WrongFormatException("Whopsie daisies! I don't understand that format!");
-            }
-            this.description = description;
-        }
-
-        public TodoTask(boolean isDone, String description) {
-            this.isDone = isDone;
-            this.description = description;
-        }
-
-        @Override
-        protected String getTaskTypeString() {
-            return squareBracketWrapper("T");
-        }
-
-        @Override
-        protected String saveToFileString() {
-            return "TODO | " + (isDone ? "1" : "0") + " | " + description;
-        }
-
-        @Override
-        protected String getDescription(String input) {
-            if (input.split(" ", 2).length == 1) {
-                return null;
-            }
-            return input.split(" ", 2)[1];
-        }
-
-        @Override
-        public String toString() {
-            return getTaskTypeString() + squareBracketWrapper(isDone ? "X" : " ") + " " + description;
-        }
-    }
-
-    private static final class DeadlineTask extends Task {
-        private LocalDate dateEnd;
-
-        public DeadlineTask(String task) throws Duke.WrongFormatException {
-            String description = getDescription(task);
-            if (description == null) {
-                throw new Duke.WrongFormatException("Whopsie daisies! I don't understand that format!");
-            }
-            this.description = description;
-        }
-
-        public DeadlineTask(boolean isDone, String description, String dateEnd) {
-            this.isDone = isDone;
-            this.description = description;
-            this.dateEnd = LocalDate.parse(dateEnd);
-        }
-
-        @Override
-        protected String getTaskTypeString() {
-            return squareBracketWrapper("D");
-        }
-
-        @Override
-        protected String saveToFileString() {
-            return "DEADLINE | " + (isDone ? "1" : "0") + " | " + description + " | " + dateEnd;
-        }
-
-        @Override
-        protected String getDescription(String input) {
-            if (input.split(" ", 2).length == 1) {
-                return null;
-            }
-
-            String[] split = input.split(" ", 2)[1].split(" /by ");
-
-            if (split.length == 1) {
-                return null;
-            }
-
-            try {
-                this.dateEnd = LocalDate.parse(split[1]);
-            } catch (DateTimeException e) {
-                return null;
-            }
-
-            return split[0];
-        }
-
-        @Override
-        public String toString() {
-            return getTaskTypeString() + squareBracketWrapper(isDone ? "X" : " ") + " " + description
-                    + " (by: " + dateEnd.format(formatter) + ")";
-        }
-    }
-
-    private static final class EventTask extends Task {
-        private LocalDate dateStart;
-        private LocalDate dateEnd;
-
-        public EventTask(String task) throws Duke.WrongFormatException {
-            String description = getDescription(task);
-            if (description == null) {
-                throw new Duke.WrongFormatException("Whopsie daisies! I don't understand that format!");
-            }
-            this.description = description;
-        }
-
-        public EventTask(boolean isDone, String description, String dateStart, String dateEnd) {
-            this.isDone = isDone;
-            this.description = description;
-            this.dateStart = LocalDate.parse(dateStart);
-            this.dateEnd = LocalDate.parse(dateEnd);
-        }
-
-        @Override
-        protected String getTaskTypeString() {
-            return squareBracketWrapper("E");
-        }
-
-        @Override
-        protected String saveToFileString() {
-            return "EVENT | " + (isDone ? "1" : "0") + " | " + description + " | " + dateStart + " | " + dateEnd;
-        }
-
-        @Override
-        protected String getDescription(String input) {
-            if (input.split(" ", 2).length == 1) {
-                return null;
-            }
-
-            String[] split = input.split(" ", 2)[1].split(" /from ");
-            if (split.length == 1) {
-                return null;
-            }
-
-            String[] split2 = split[1].split(" /to ");
-            if (split2.length == 1) {
-                return null;
-            }
-
-            try {
-                this.dateStart = LocalDate.parse(split2[0]);
-                this.dateEnd = LocalDate.parse(split2[1]);
-            } catch (DateTimeException e) {
-                return null;
-            }
-
-            return split[0];
-        }
-
-        @Override
-        public String toString() {
-            return getTaskTypeString() + squareBracketWrapper(isDone ? "X" : " ") + " " + description
-                    + " (from: " + dateStart.format(formatter) + " to: " + dateEnd.format(formatter) + ")";
-        }
     }
 }
